Revised Calibration Behavior
The settings now provide a pivot point between the upper and lower bound,
which the calibration phase attempts to nudge you towards gently.  This
is accomplished by lying about the current target by one in whichever
direction you need to be moving.

This also changes the calibration to weight recent readings higher than
past readings when determining targets, so that the target doesn't lock
up after holding steady for a while.  Judging from my last workout,
this probably would benefit from a rolling window, rather than a weighted
average.

Finally, this change updates the bpm debug mode to visualze this
information, and where the rolling bpm is relative to the range limits
and pivot.